+++ ios/PacketTunnelCore/Actor/ProtocolObfuscator.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import MullvadLogging
 import MullvadREST
 import MullvadRustRuntime
 import MullvadSettings
@@ -19,14 +18,13 @@ public struct ProtocolObfuscationResult {
 }
 
 public protocol ProtocolObfuscation {
-    func obfuscate(_ endpoint: SelectedEndpoint) -> ProtocolObfuscationResult
+    func obfuscate(_ endpoint: SelectedEndpoint, clientPublicKey: PublicKey) -> ProtocolObfuscationResult
     var transportLayer: TransportLayer? { get }
     var remotePort: UInt16 { get }
 }
 
 public class ProtocolObfuscator<Obfuscator: TunnelObfuscation>: ProtocolObfuscation {
     var tunnelObfuscator: TunnelObfuscation?
-    var logger = Logger(label: "ProtocolObfuscator")
 
     public init() {}
 
@@ -44,7 +42,7 @@ public class ProtocolObfuscator<Obfuscator: TunnelObfuscation>: ProtocolObfuscat
     ///
     /// Note: Obfuscation currently only supports IPv4. If the endpoint uses IPv6,
     /// obfuscation is skipped and the endpoint is returned as-is with obfuscation disabled.
-    public func obfuscate(_ endpoint: SelectedEndpoint) -> ProtocolObfuscationResult {
+    public func obfuscate(_ endpoint: SelectedEndpoint, clientPublicKey: PublicKey) -> ProtocolObfuscationResult {
         remotePort = endpoint.socketAddress.port
 
         // Extract obfuscation protocol from the bundled obfuscation method
@@ -59,7 +57,11 @@ public class ProtocolObfuscator<Obfuscator: TunnelObfuscation>: ProtocolObfuscat
             case let .quic(hostname, token):
                 .quic(hostname: hostname, token: token)
             case .lwo:
-                .lwo
+                if let key = PublicKey(rawValue: endpoint.publicKey) {
+                    .lwo(serverPublicKey: key)
+                } else {
+                    nil
+                }
             }
 
         // If obfuscation is disabled, return endpoint as-is
@@ -68,18 +70,11 @@ public class ProtocolObfuscator<Obfuscator: TunnelObfuscation>: ProtocolObfuscat
             return .init(endpoint: endpoint)
         }
 
-        guard let publicKey = PublicKey(rawValue: endpoint.publicKey) else {
-            logger.error("Could not create public key from endpoint data")
-
-            tunnelObfuscator = nil
-            return .init(endpoint: endpoint)
-        }
-
         let obfuscator = Obfuscator(
             remoteAddress: endpoint.socketAddress.ip,
-            tcpPort: remotePort,
+            remotePort: remotePort,
             obfuscationProtocol: obfuscationProtocol,
-            clientPublicKey: publicKey
+            clientPublicKey: clientPublicKey
         )
 
         obfuscator.start()